Доделки в первом, а так же второе и третье

+++ cs/TagsCloudVisualization/Spiral.cs
@@ -1,16 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 
 namespace TagsCloudVisualization
 {
     public class Spiral
     {
-        private readonly int parameter;
+        private readonly double parameter;
         private readonly double stepInRadians;
         private double phi;
 
-        public Spiral(int parameter, int stepInDegrees)
+        public Spiral(double parameter, int stepInDegrees)
         {
             this.parameter = parameter;
             stepInRadians = stepInDegrees * Math.PI / 180;
